Update opentx.h
Wanted, only as a personal preference, to have the final Status LED indicator match my red gimbals on my Radiomaster Boxer. Have never liked it since it was changed to green in a recent update. Brand new to this, so likely i'm not following 'protocol'. But, we never learn if we don't try. I have been trying to create a fork, but kept going round in circles. Any easy to follow advice is welcomed on that front. I'm not bothered about requesting a wholesale change for everyone, simply my own, and some friends' radios. Thanks